% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluey-package.R
\docType{package}
\name{gluey-package}
\alias{gluey-package}
\title{gluey: Advanced cli-like markdown glue syntax}
\description{
Generate markdown text using the familiar syntax of 'glue' with enhanced templating features. Unlike 'cli' which produces styled console output, 'gluey' outputs plain markdown text, making it ideal for document generation in R Markdown and Quarto documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vorpalvorpal/gluey}
  \item Report bugs at \url{https://github.com/vorpalvorpal/gluey/issues}
}

}
\author{
\strong{Maintainer}: Robin Shannon \email{32329438+vorpalvorpal@users.noreply.github.com}

}
\keyword{internal}
