% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluey.R
\name{knit_print.gluey}
\alias{knit_print.gluey}
\title{Knit print method for gluey objects}
\usage{
knit_print.gluey(x, ...)
}
\arguments{
\item{x}{A gluey object}

\item{...}{Additional arguments passed to knitr::asis_output}
}
\value{
The result of knitr::asis_output
}
\description{
Knit print method for gluey objects
}
