% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralise.R
\name{no}
\alias{no}
\title{Format numeric zero as "no"}
\usage{
no(expr)
}
\arguments{
\item{expr}{Expression that evaluates to numeric}
}
\value{
The value with special formatting for zero
}
\description{
Format numeric zero as "no"
}
\examples{
nfile <- 0
gluey("Found {no(nfile)} file{?s}.")
# "Found no files."

nfile <- 1
gluey("Found {no(nfile)} file{?s}.")
# "Found 1 file."
}
