% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluey.R
\name{print.gluey}
\alias{print.gluey}
\title{Print method for gluey objects}
\usage{
\method{print}{gluey}(x, ...)
}
\arguments{
\item{x}{A gluey object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for gluey objects
}
