% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralise.R
\name{qty}
\alias{qty}
\title{Set quantity for pluralization without displaying it}
\usage{
qty(expr)
}
\arguments{
\item{expr}{Expression that sets the quantity}
}
\value{
Empty string
}
\description{
Set quantity for pluralization without displaying it
}
\examples{
nupd <- 3
ntotal <- 10
gluey("{nupd}/{ntotal} {qty(nupd)} file{?s} {?needs/need} updates")
# "3/10 files need updates"
}
