% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluey.R
\docType{package}
\name{gluey-package}
\alias{gluey-package}
\title{Gluey: Markdown Text Generation with Enhanced Templates}
\description{
Generate markdown text using familiar glue syntax with enhanced
templating features and pluralization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vorpalvorpal/gluey}
  \item Report bugs at \url{https://github.com/vorpalvorpal/gluey/issues}
}

}
\author{
\strong{Maintainer}: Robin Shannon \email{first.last@example.com}

}
